requirements.txt.txt
tensorflow==2.10.0
flask==3.1.1
streamlit
onnxruntime
pillow
numpy
